+++ generators/app/templates/gulp/tasks/postcss.js
@@ -4,30 +4,32 @@ import path from 'path';
 import autoprefixer from 'autoprefixer';
 import gulpif from 'gulp-if';
 import gulp from 'gulp';
+import postCssImport from 'postcss-import';
 import { plugins, args, config, taskTarget, browserSync } from '../utils';
 
 let dirs = config.directories;
 let entries = config.entries;
 let dest = path.join(taskTarget, dirs.styles.replace(/^_/, ''));
+var postCssPlugins = [
+  postCssImport({
+    path: [
+      path.join(dirs.source, dirs.styles),
+      path.join(dirs.source, dirs.modules)
+    ]
+  }),
+  autoprefixer()
+];
 
-// Less compilation
-gulp.task('less', () => {
+// PostCSS compilation
+gulp.task('postcss', () => {
   return gulp
     .src(entries.css, { cwd: path.join(dirs.source, dirs.styles) })
     .pipe(plugins.plumber())
     .pipe(gulpif(!args.production, plugins.sourcemaps.init({ loadMaps: true })))
-    .pipe(
-      plugins.less({
-        paths: [
-          path.join(dirs.source, dirs.styles),
-          path.join(dirs.source, dirs.modules)
-        ]
-      })
-    )
+    .pipe(plugins.postcss(postCssPlugins))
     .on('error', function(err) {
       plugins.util.log(err);
     })
-    .pipe(plugins.postcss([autoprefixer()]))
     .pipe(
       plugins.rename(function(path) {
         // Remove 'source' directory as well as prefixed folder underscores